Create findHighestPopulation.js
This file prints out the years that achieved the maximum population.  The input is an object of object each of which represents a person and details their birth and death years.   The data set hard coded into the file is.

                     {person1 : {"birthYear" : 1981, "deathYear": 1999}, 
                      person2 : {birthYear : 1920, "deathYear" : 1991},
                      person3 : {birthYear : 1901, "deathYear" : 1956},
                      person4 : {birthYear : 1973, "deathYear" : 2000},
                      person5 : {birthYear : 1946, "deathYear" : 1983},
                      person6 : {birthYear : 1955, "deathYear" : 1960},
                      person7 : {birthYear : 1963, "deathYear" : 1992},
                      person8 : {birthYear : 1982, "deathYear" : 1997},
                      person9 : {birthYear : 1949, "deathYear" : 1988},
                      person10 : {birthYear : 1913, "deathYear" : 1965},
                      person11 : {birthYear : 1922, "deathYear" : 1979}};

The expected outcome is a line printed to the console that gives the maximum population achieved and the year(s) it was achieved.  For the above data set the expected outcome is "Highest population achieved was 7 during the year(s)  1955 1956 1982 1983".